+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 
 * handling optional position properties in RVDAS provider
   [#168](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/168)
+* handling connection errors better in RVDAS provider
+  [#167](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/167)
 
 ## [0.7.0] - 2025-06-14